+++ src/AppBundle/Controller/ReservaController.php
@@ -22,11 +22,22 @@ class ReservaController extends Controller
      * @Route("/cliente/reserva", name="reserva_cliente")
      */
     public function reservaClienteAction(Request $request){
-        $idr=$this->get('session')->get("id");
-        $em = $this->getDoctrine()->getEntityManager();
-        $reservas = $em->getRepository('AppBundle:reserva')->findBy(array('idUsuario'=>$idr));
 
-        return $this->render('reserva/reserva_cliente.html.twig',['listado'=>$reservas]);
+      $session = $request->getSession();
+
+        if($session->has("id")){
+
+            $idr=$this->get('session')->get("id");
+            $em = $this->getDoctrine()->getEntityManager();
+            $reservas = $em->getRepository('AppBundle:reserva')->findBy(array('idUsuario'=>$idr));
+            return $this->render('reserva/reserva_cliente.html.twig',['listado'=>$reservas]);
+
+        }else{
+
+            $this->get('session')->getFlashBag()->add('mensaje','Debe estar Logueado para este contenido');
+            return $this->redirect($this->generateUrl('loguear'));
+        }
+        
     }
 
      /**
@@ -34,42 +45,53 @@ class ReservaController extends Controller
     */
     public function reservaCreateAction(Request $request)
     {
-      //consulta las servicios para crear una lista que se usará para cargar en el formulario (ChoiceType)
-      $listado_servicio=$this->getDoctrine()->getRepository('AppBundle:reserva_tipo_servicio')->findAll();
-      foreach($listado_servicio as $ob){
-        $lista[$ob->getId()]=$ob->getNombre();
-      }
+
+      $session = $request->getSession();
+
+      if($session->has("id")){
+             //consulta las servicios para crear una lista que se usará para cargar en el formulario (ChoiceType)
+          $listado_servicio=$this->getDoctrine()->getRepository('AppBundle:reserva_tipo_servicio')->findAll();
+          foreach($listado_servicio as $ob){
+            $lista[$ob->getId()]=$ob->getNombre();
+            }
 
       //se crea el formulario
-      $rese=new reserva;
-      $form=$this->createFormBuilder($rese)
-      ->add('reservaFecha',DateType::class,array('label'=>'Fecha de reserva','attr' => array('class'=>'form-control','style'=>'margin-bottom:15px')))
-      ->add('reservaLugar',TextType::class,array('label'=>'Lugar','attr' => array('class'=>'form-control','style'=>'margin-bottom:15px')))
-      ->add('reservaTipoId',ChoiceType::class,array('label'=>'Seleccione tipo de Servicio','placeholder'=>false,
-      'choices'=>array_flip($lista),
-      'attr' => array('class'=>'form-control','style'=>'margin-bottom:15px')))
-      ->add('Guardar',SubmitType::class,array('attr' => array('class'=>'btn btn-success btn-sm','style'=>'margin-bottom:15px')))
-      ->getForm();
-      $form->handleRequest($request);
+            $rese=new reserva;
+            $form=$this->createFormBuilder($rese)
+            ->add('reservaFecha',DateType::class,array('label'=>'Fecha de reserva','attr' => array('class'=>'form-control','style'=>'margin-bottom:15px')))
+            ->add('reservaLugar',TextType::class,array('label'=>'Lugar','attr' => array('class'=>'form-control','style'=>'margin-bottom:15px')))
+            ->add('reservaTipoId',ChoiceType::class,array('label'=>'Seleccione tipo de Servicio','placeholder'=>false,
+            'choices'=>array_flip($lista),
+            'attr' => array('class'=>'form-control','style'=>'margin-bottom:15px')))
+            ->add('Guardar',SubmitType::class,array('attr' => array('class'=>'btn btn-success btn-sm','style'=>'margin-bottom:15px')))
+            ->getForm();
+            $form->handleRequest($request);
 
       //Después de llenar el formulario
-       if ($form->isSubmitted() && $form->isValid()) {
-         $rese->setReservaFecha($form['reservaFecha']->getData());
-         $rese->setReservaLugar($form['reservaLugar']->getData());
-         $cat = $this->getDoctrine()->getRepository(reserva_tipo_servicio::class)->find($form['reservaTipoId']->getData());
-         $rese->setreservaTipoId($cat);
-         $idr=$this->get('session')->get("id");
-         $rese->setidUsuario($idr);
+            if ($form->isSubmitted() && $form->isValid()) {
+             $rese->setReservaFecha($form['reservaFecha']->getData());
+             $rese->setReservaLugar($form['reservaLugar']->getData());
+             $cat = $this->getDoctrine()->getRepository(reserva_tipo_servicio::class)->find($form['reservaTipoId']->getData());
+             $rese->setreservaTipoId($cat);
+             $idr=$this->get('session')->get("id");
+             $rese->setidUsuario($idr);
 
          /** @var Symfony\Component\HttpFoundation\File\UploadedFile $file */
-         $em=$this->getDoctrine()->getManager();
-         $em->persist($rese);
-         $em->flush();
-         $this->addFlash('notice','Reserva Agregada con éxito');
-        return $this->redirectToRoute('reserva_cliente');
-       }//fin
+             $em=$this->getDoctrine()->getManager();
+             $em->persist($rese);
+             $em->flush();
+             $this->addFlash('notice','Reserva Agregada con éxito');
+             return $this->redirectToRoute('reserva_cliente');
+             }//fin
 
-      return $this->render('reserva/agregar_reserva.html.twig',array('form' => $form->createView()));
+            return $this->render('reserva/agregar_reserva.html.twig',array('form' => $form->createView()));
+
+        }else{
+
+            $this->get('session')->getFlashBag()->add('mensaje','Debe estar Logueado para este contenido');
+            return $this->redirect($this->generateUrl('loguear'));
+        }
+     
 
     }
 
@@ -78,23 +100,35 @@ class ReservaController extends Controller
      */
     public function tipoReservaAction(Request $request){
 
-    	if($request->getMethod()=="POST"){
+      $session = $request->getSession();
 
-    		$post = new reserva_tipo_servicio();
-    		$nombre = $request->get('nombre');
-    		$descripcion = $request->get('descripcion');
-    		$precio = $request->get('precio');
+      if($session->has("id")){
 
-    		$post->setNombre($nombre);
-    		$post->setDescripcion($descripcion);
-    		$post->setPrecio($precio);
-    		$em = $this->getDoctrine()->getEntityManager();
-            $em->persist($post);
-            $flush = $em->flush();
+        if($request->getMethod()=="POST"){
 
-            return $this->render('reserva/tipo_reserva.html.twig');
-    	}
-    	return $this->render('reserva/tipo_reserva.html.twig');
+          $post = new reserva_tipo_servicio();
+          $nombre = $request->get('nombre');
+          $descripcion = $request->get('descripcion');
+          $precio = $request->get('precio');
+
+          $post->setNombre($nombre);
+          $post->setDescripcion($descripcion);
+          $post->setPrecio($precio);
+          $em = $this->getDoctrine()->getEntityManager();
+          $em->persist($post);
+          $flush = $em->flush();
+
+          return $this->render('reserva/tipo_reserva.html.twig');
+        }
+        return $this->render('reserva/tipo_reserva.html.twig');
+
+      }else{
+
+          $this->get('session')->getFlashBag()->add('mensaje','Debe estar Logueado para este contenido');
+          return $this->redirect($this->generateUrl('loguear'));
+      }
+
+    	
     }
 
 
@@ -102,10 +136,22 @@ class ReservaController extends Controller
      * @Route("/admi/listado_tipo_reserva", name="reserva_tipoList")
      */
     public function listTporeservAction(Request $request){
-    	$em = $this->getDoctrine()->getEntityManager();
-    	$listado = $em->getRepository('AppBundle:reserva_tipo_servicio')->findAll();
 
-    	return $this->render('reserva/listado_tipo_reserva.html.twig',['listado'=>$listado]);
+      $session = $request->getSession();
+
+        if($session->has("id")){
+          $em = $this->getDoctrine()->getEntityManager();
+          $listado = $em->getRepository('AppBundle:reserva_tipo_servicio')->findAll();
+
+          return $this->render('reserva/listado_tipo_reserva.html.twig',['listado'=>$listado]);
+
+        }else{
+
+            $this->get('session')->getFlashBag()->add('mensaje','Debe estar Logueado para este contenido');
+            return $this->redirect($this->generateUrl('loguear'));
+        }
+
+    	
     }
 
      /**
@@ -113,12 +159,23 @@ class ReservaController extends Controller
      */
     public function eliminarTipoReservaAction(Request $request,$id){
 
-    	$em = $this->getDoctrine()->getEntityManager();
-        $tipo = $em->getRepository('AppBundle:reserva_tipo_servicio')->find($id);
-        $em->remove($tipo);
-        $em->flush();
+      $session = $request->getSession();
 
-        return $this->redirect($this->generateUrl('reserva_tipo'));
+        if($session->has("id")){
+          $em = $this->getDoctrine()->getEntityManager();
+          $tipo = $em->getRepository('AppBundle:reserva_tipo_servicio')->find($id);
+          $em->remove($tipo);
+          $em->flush();
+
+          return $this->redirect($this->generateUrl('reserva_tipo'));
+
+        }else{
+
+            $this->get('session')->getFlashBag()->add('mensaje','Debe estar Logueado para este contenido');
+            return $this->redirect($this->generateUrl('loguear'));
+        }     
+
+    	
     }
 
 
@@ -126,9 +183,20 @@ class ReservaController extends Controller
      * @Route("/admi/ver_tipo/{id}", name="verTipoReserva")
      */
     public function verTipoReserva(Request $request,$id){
-    	$em = $this->getDoctrine()->getEntityManager();
-        $tipo = $em->getRepository('AppBundle:reserva_tipo_servicio')->find($id);
+      $session = $request->getSession();
 
-        return $this->render('reserva/ver_tipoReserva.html.twig',['tipo_ver'=>$tipo]);
+        if($session->has("id")){
+          $em = $this->getDoctrine()->getEntityManager();
+          $tipo = $em->getRepository('AppBundle:reserva_tipo_servicio')->find($id);
+
+          return $this->render('reserva/ver_tipoReserva.html.twig',['tipo_ver'=>$tipo]);
+
+        }else{
+
+            $this->get('session')->getFlashBag()->add('mensaje','Debe estar Logueado para este contenido');
+            return $this->redirect($this->generateUrl('loguear'));
+        }
+
+    	  
     }
 }
